+++ cmd/e2e_test.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
-	_ "github.com/redhat-appstudio/e2e-tests/pkg/framework/application-service"
-	_ "github.com/redhat-appstudio/e2e-tests/pkg/framework/common"
+	_ "github.com/redhat-appstudio/e2e-tests/pkg/tests/common"
+	_ "github.com/redhat-appstudio/e2e-tests/pkg/tests/has"
 )
 
 var _ = ginkgo.SynchronizedBeforeSuite(func() []byte {
