--- pkg/framework/application-service/controller/controller.go
@@ -1,21 +1,0 @@
-package controller
-
-import (
-	"fmt"
-
-	"github.com/redhat-appstudio/e2e-tests/pkg/client"
-)
-
-type HASSuiteController struct {
-	*client.K8sClient
-}
-
-func NewHasSuiteController() (*HASSuiteController, error) {
-	client, err := client.NewK8SClient()
-	if err != nil {
-		return nil, fmt.Errorf("Error creating client-go")
-	}
-	return &HASSuiteController{
-		client,
-	}, nil
-}
