--- pkg/framework/common/controller/pipelines.go
@@ -1,22 +1,0 @@
-package controller
-
-import (
-	"context"
-
-	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
-	"k8s.io/apimachinery/pkg/types"
-)
-
-// GetClusterTask return a clustertask object from cluster and if don't exist returns an error
-func (h *CommonSuiteController) GetClusterTask(name string, namespace string) (*v1beta1.ClusterTask, error) {
-	namespacedName := types.NamespacedName{
-		Name:      name,
-		Namespace: namespace,
-	}
-	clusterTask := &v1beta1.ClusterTask{}
-
-	if err := h.KubeRest().Get(context.TODO(), namespacedName, clusterTask); err != nil {
-		return nil, err
-	}
-	return clusterTask, nil
-}
